// src/lib/types/show.type.ts
export interface IShowItem {
  backdrop_path: string
  first_air_date: string
  genre_ids: number
  id: number
  name: string
  origin_country: string[]
  original_language: string
  original_name: string
  overview: string
  popularity: number
  poster_path: string
  vote_average: number
  vote_count: number
}

export type ShowType = IShowItem[]

export interface ShowResponse {
  page: number
  results: ShowType
  total_pages: number
  total_results: number
}